requirements.txt
mysql-connector-python
pillow
python-Levenshtein
pandas